requirements.txt
cmake
lit
timm==0.9.2
opencv-python
tqdm
pyyaml
joblib
glog
scikit-image
tensorboardX
albumentations==1.1.0
einops
imgaug  # required by albumentations[imgaug]
numpy<2
mamba-ssm==2.2.2
causal-conv1d==1.4.0
setuptools      
wheel         
torch           
torchvision   
pillow      
scipy      
imageio      
pyiqa
pandas
PyWavelets 
bm3d    
piq